Add a max_connections transport option.
Limits the number of parallel requests processed at once.
Waiting requests are kept in the accept queue.
The limit is not explicitly observed, but it should be
around the given value at any time. Defaults to 1024.

Thanks to ostinelli's benchmark to point out the issue hopefully
solved by this and the previous commit.