Add built in cowboy_http_static handler
The aim of the handler is to provide the bare minimum functionality needed for
small or temporary environments where setting up a second web server just to
serve static content is infeasible.

The handler must be configured with a document root (the 'directory' option).
The directory option can be set to an absolute or relative path in the form of
a list or binary representation of a string. The list of binary path segments
used throughout cowboy is also a valid path.

To simplify using this handler to serve files included in a release a path of
the form {priv_dir, Application, RelPath} is also supported. This will result
in the files being served out of the RelPath directory within the priv/
directory of the application.

The handler can be configured with a list of known mime types to override
the default mime type used for static resources. A list of file extensions to
mime types pairs must be provided.

These changes also include the new cowboy_http_req:set_resp_body_fun/2 function.
This function is necessary to support the file:sendfile/2 call introduced in
R15B. It's likely to be useful for other cases where the size of a response body
is known but is too large to store in an intermidiate buffer before sending the
reply. The set_resp_body_fun/2 function is supported in rest by returning
{'stream', Length, Fun} from a content type function.