Localize error reporting for rest_init/2
Previous to this patch errors that occured in handling the service
request where reported as errors in the rest_init/2 function of the
handler. This patch localizes that error messages such that only errors
that actually happen in rest_init/2 are reported as happening in
rest_init/2. If errors happen at the server level a normal crash should
occur.