Add an RFC7230 HTTP/1.1 server specification document
This document lists all rules that Cowboy HTTP server will follow
after careful analysis of the RFC7230.

Do note that not all these rules are implemented or tested yet,
and that at the time of commit there may be rules that Cowboy will
ultimately not implement (for various reasons including performance
or leaving certain aspects up to the user code).