keepalive on HEAD response without payload headers
According to RFC 7231 Section 4.3.2:

    The server SHOULD send the same header fields in response to a HEAD
    request as it would have sent if the request had been a GET, except
    that the payload header fields (Section 3.3) MAY be omitted.

Section 3.3 defines the payload header fields as:

- Content-Length
- Content-Range
- Trailer
- Transfer-Encoding

Accordingly, treating a close-delimited response as terminating when the
connection is closed (item 7 in RFC 7230 section 3.3.3) is not necessary
when the response is to a HEAD request.

This patch allows to override the connection-closing via the headers
submitted by the handler when the response is to a HEAD request. For
backwards-compatibility reasons, the default value remains 'closed'.