Improved cowboy_rest charset matching
If we match a charset from the accept header we skip the
charsets_provided callback.
If we match a charset in the accept header against a '*' then we use
that against the charsets_provided callback instead of the
accept-charsets header.
Also changed the auto appending of charsets to only occur if its a text
media type and there is not already a charset in the params.